feat: 🎸 split notfications code from functions folder
Move the notifications code into separate folder and import it to
functions. We do this to enable different domains to be as separate as
possible from firebase cloud functions

BREAKING CHANGE: 🧨 functions can't be deployed as before
Closes #157
