+++ firebase_serve.sh
@@ -6,7 +6,7 @@ command -v firebase >/dev/null 2>&1 || { echo >&2 "firebase needs to be installe
 
 
 cd client/ && npm run build-dev && cd ..
-cd server/functions && npm run build && cd ../../
+cd server/notifications && npm run build && cd ../../
 cd server/data-store && npm run tsc-build && cd ../../
 
 firebase serve
