+++ install.sh
@@ -5,4 +5,5 @@ command -v npm >/dev/null 2>&1 || { echo >&2 "npm needs to be installed"; exit 1
 
 cd client/ && npm install && cd ..
 cd server/functions && npm install && cd ../../
+cd server/notifications && npm install && cd ../../
 cd server/data-store && npm install && cd ../../
