+++ server/functions/package.json
@@ -1,10 +1,8 @@
 {
   "name": "functions",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
+    "serve": "firebase serve --only functions",
+    "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
@@ -12,18 +10,16 @@
   "engines": {
     "node": "10"
   },
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "dependencies": {
     "@google-cloud/pubsub": "^1.7.0",
+    "express": "^4.17.1",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.0",
     "grpc": "^1.24.2"
   },
   "devDependencies": {
-    "@types/node": "^13.11.0",
-    "firebase-functions-test": "^0.1.6",
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2"
+    "firebase-functions-test": "^0.1.6"
   },
   "private": true
 }
